final v1

+++ screens/user/UserProductsScreen.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState, useCallback } from "react";
-import { FlatList, Button, Alert, StyleSheet } from "react-native";
+import { View, Text, FlatList, Button, Alert, StyleSheet } from "react-native";
 import { useSelector, useDispatch } from "react-redux";
 import { HeaderButtons, Item } from "react-navigation-header-buttons";
 import HeaderButton from "../../components/UI/HeaderButton";
@@ -20,24 +20,39 @@ const UserProductsScreen = (props) => {
   const deleteHandler = async (id) => {
     setError(null);
     try {
-      const rmProduct = await dispatch(deleteProduct(id));
+      //const rmProduct = await dispatch(deleteProduct(id));
+      Alert.alert("Are you sure?", "Do you want to delete the product?", [
+        {
+          text: "No",
+          style: "default",
+          onPress: () => {
+            return;
+          },
+        },
+        {
+          text: "Yes",
+          style: "destructive",
+          onPress: async () => {
+            await dispatch(deleteProduct(id));
+          },
+        },
+      ]);
     } catch (err) {
       setError(err.message);
     }
-    Alert.alert("Are you sure?", "Do you want to delete the product?", [
-      { text: "No", style: "default" },
-      {
-        text: "Yes",
-        style: "destructive",
-        onPress: () => rmProduct,
-      },
-    ]);
   };
   useEffect(() => {
     if (error) {
       Alert.alert("Alert!", error, [{ text: "Okay" }]);
     }
   }, [error]);
+  if (userProducts.length === 0) {
+    return (
+      <View style={styles.noProducts}>
+        <Text>No specific user products. Create some products</Text>
+      </View>
+    );
+  }
   return (
     <FlatList
       data={userProducts}
@@ -63,7 +78,15 @@ const UserProductsScreen = (props) => {
   );
 };
 
-const styles = StyleSheet.create({});
+const styles = StyleSheet.create({
+  noProducts: {
+    flex: 1,
+    justifyContent: "center",
+    alignItems: "center",
+    fontFamily: "popins-medium",
+    fontSize: 14,
+  },
+});
 
 UserProductsScreen.navigationOptions = (navData) => {
   return {
